Indentation

+++ xxGraphicD3D6.cpp
@@ -8,8 +8,8 @@
 #include "xxGraphicD3D.h"
 #include "xxGraphicD3D6.h"
 
-#define DIRECTDRAW_VERSION      0x600
-#define DIRECT3D_VERSION        0x600
+#define DIRECTDRAW_VERSION          0x600
+#define DIRECT3D_VERSION            0x600
 #include "dxsdk/ddraw.h"
 #include "dxsdk/d3d.h"
 interface DECLSPEC_UUID("9c59509a-39bd-11d1-8c4a-00c04fd930c5") IDirectDraw4;
@@ -17,14 +17,14 @@ interface DECLSPEC_UUID("0B2B8630-AD35-11D0-8EA6-00609797EA5B") IDirectDrawSurfa
 interface DECLSPEC_UUID("bb223240-e72b-11d0-a9b4-00aa00c0993e") IDirect3D3;
 interface DECLSPEC_UUID("84E63dE0-46AA-11CF-816F-0000C020156E") IDirect3DHALDevice;
 interface DECLSPEC_UUID("93281502-8cf8-11d0-89ab-00a0c9054129") IDirect3DTexture2;
-typedef HRESULT (WINAPI * PFN_DIRECT_DRAW_CREATE)(GUID*, LPDIRECTDRAW*, IUnknown*);
-#define D3DRTYPE_CONSTANTBUFFER 0
-#define D3DRTYPE_INDEXBUFFER    1
-#define D3DRTYPE_VERTEXBUFFER   2
+typedef HRESULT (WINAPI *PFN_DIRECT_DRAW_CREATE)(GUID*, LPDIRECTDRAW*, IUnknown*);
+#define D3DRTYPE_CONSTANTBUFFER     0
+#define D3DRTYPE_INDEXBUFFER        1
+#define D3DRTYPE_VERTEXBUFFER       2
 
-static HMODULE                  g_ddrawLibrary = nullptr;
-static LPDIRECTDRAW4            g_ddraw = nullptr;
-static LPDIRECTDRAWSURFACE4     g_primarySurface = nullptr;
+static HMODULE                      g_ddrawLibrary = nullptr;
+static LPDIRECTDRAW4                g_ddraw = nullptr;
+static LPDIRECTDRAWSURFACE4         g_primarySurface = nullptr;
 
 //==============================================================================
 //  Resource Type
@@ -944,7 +944,7 @@ void xxDrawIndexedD3D6(uint64_t commandEncoder, uint64_t indexBuffer, int indexC
     }
     for (int i = 0; i < tailIndexCount; ++i)
     {
-        (*q++) = (*p++) + vertexOffset;
+        (*q++) = (WORD)((*p++) + vertexOffset);
     }
 
     d3dDevice->DrawIndexedPrimitiveVB(D3DPT_TRIANGLELIST, vertexBuffer, wordIndexBuffer, indexCount, 0);
